requirements.txt
# Core dependencies for Career AI Assistant
gradio>=4.0.0
google-generativeai>=0.3.0
redis>=4.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
asyncio

# Email and Communication
sendgrid>=6.9.0

# ML and AI
numpy>=1.24.0
rapidfuzz>=3.0.0

# Vector search
pinecone-client>=2.2.0

# ONNX runtime (for fallback)
onnxruntime>=1.15.0

# Monitoring and logging
python-json-logger>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Additional utilities (from old requirements)
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0
plotly>=5.17.0
pandas>=2.0.0